added changes for oclif
