+++ tsconfig.json
@@ -1,10 +1,16 @@
 {
-    "extends": "@tsconfig/recommended/tsconfig.json",
-    "files": [
-        "./cli.ts",
-        "./index.ts"
-    ],
     "compilerOptions": {
-        "outDir": "./dist"
-    }
-}
+      "declaration": true,
+      "importHelpers": true,
+      "esModuleInterop": true,
+      "module": "commonjs",
+      "outDir": "lib",
+      "rootDir": "src",
+      "strict": true,
+      "target": "es2017"
+    },
+    "include": [
+      "src/**/*"
+    ]
+  }
+
